rename geo_cslc to s1_geocode_slc
+ rename geo_cslc.py to s1_geocode_slc.py and update its usage

+ rename geo_cslc_metadata to geo_metadata

+ fix typos in journal.info() calls by adding the missing "s1_" prefix
